requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
pydantic>=2.7.0

# Ekstrakcja tekstu z plików
pdfplumber>=0.11.0
python-docx>=1.1.0

# Embeddingi semantyczne
sentence-transformers>=3.0.0
numpy>=1.26.0
